fix: do not create analysisruns with initial deploy (#722)

+++ test/fixtures/e2e_suite.go
@@ -28,11 +28,21 @@ import (
 	istioutil "github.com/argoproj/argo-rollouts/utils/istio"
 )
 
+const (
+	// E2E_INSTANCE_ID is the instance id label attached to objects created by the e2e tests
+	EnvVarE2EInstanceID = "E2E_INSTANCE_ID"
+	// E2E_WAIT_TIMEOUT is a timeout in seconds when waiting for a test condition (default: 60)
+	EnvVarE2EWaitTimeout = "E2E_WAIT_TIMEOUT"
+	// E2E_POD_DELAY slows down pod startup and shutdown by the value in seconds (default: 0)
+	// Used humans slow down rollout activity during a test
+	EnvVarE2EPodDelay = "E2E_POD_DELAY"
+	// E2E_DEBUG makes e2e testing easier to debug by not tearing down the suite
+	EnvVarE2EDebug = "E2E_DEBUG"
+)
+
 var (
-	// E2EWaitTimeout is the timeout when waiting for a condition to become true
 	E2EWaitTimeout time.Duration = time.Second * 60
 
-	// E2E Label is the
 	E2ELabel = "argo-rollouts-e2e"
 
 	serviceGVR = schema.GroupVersionResource{
@@ -47,10 +57,10 @@ var (
 )
 
 func init() {
-	if e2elabel, ok := os.LookupEnv("E2E_INSTANCE_ID"); ok {
+	if e2elabel, ok := os.LookupEnv(EnvVarE2EInstanceID); ok {
 		E2ELabel = e2elabel
 	}
-	if e2eWaitTimeout, ok := os.LookupEnv("E2E_WAIT_TIMEOUT"); ok {
+	if e2eWaitTimeout, ok := os.LookupEnv(EnvVarE2EWaitTimeout); ok {
 		timeout, err := strconv.Atoi(e2eWaitTimeout)
 		if err != nil {
 			panic(fmt.Sprintf("Invalid wait timeout seconds: %s", e2eWaitTimeout))
@@ -91,7 +101,7 @@ func (s *E2ESuite) SetupSuite() {
 		flag.Parse()
 	}
 
-	if delayStr := os.Getenv("E2E_POD_DELAY"); delayStr != "" {
+	if delayStr := os.Getenv(EnvVarE2EPodDelay); delayStr != "" {
 		delay, err := strconv.Atoi(delayStr)
 		s.CheckError(err)
 		s.podDelay = delay
@@ -99,6 +109,10 @@ func (s *E2ESuite) SetupSuite() {
 }
 
 func (s *E2ESuite) TearDownSuite() {
+	if os.Getenv(EnvVarE2EDebug) == "true" {
+		s.log.Info("skipping resource cleanup")
+		return
+	}
 	s.DeleteResources(E2ELabel)
 }
